// src/app/page/master/unit-laporan/unit-laporan.interface.ts
import { EventEmitter } from '@angular/core';

export interface UnitLaporan {
    kode?: any;
    kdKelompokProduk?: number;
    levelUnitLaporan?: string;
    noUrut?: string;
    namaUnitLaporan?: string;
    kdUnitLaporanHead?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}